backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==25.1.0
mysqlclient==2.2.8
packaging==26.0
python-dotenv==1.2.2
sqlparse==0.5.5
typing_extensions==4.15.0
